control-center/backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
websockets==12.0
psutil==5.9.8
sqlalchemy==2.0.25
aiosqlite==0.19.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
requests>=2.32.3
slowapi==0.1.9
PyJWT>=2.10.1
bcrypt>=4.2.1
python-multipart==0.0.6
email-validator==2.1.1
httpx==0.27.0
boto3>=1.34.0
